+++ demisto_sdk/commands/common/hook_validations/pack_unique_files.py
@@ -31,7 +31,7 @@ CONTRIBUTORS_LIST = ['partner', 'developer', 'community']
 SUPPORTED_CONTRIBUTORS_LIST = ['partner', 'developer']
 ISO_TIMESTAMP_FORMAT = '%Y-%m-%dT%H:%M:%SZ'
 ALLOWED_CERTIFICATION_VALUES = ['certified', 'verified']
-SUPPORT_TYPES = ['partner', 'community', 'xsoar']
+SUPPORT_TYPES = ['community', 'xsoar'] + SUPPORTED_CONTRIBUTORS_LIST
 
 
 class PackUniqueFilesValidator(BaseValidator):